[DotNetCore] Prevent OutputType being added to library projects
Library projects do not have an OutputType property defined. When
these projects were saved the OutputType was being saved as Exe in
the project file which then results in not being able to compile
these projects. Now if the .NET Core project does not have an
OutputType property defined in the project it is removed when
saving the project file.