[DotNetCore] Fix package dependencies when multi-targeting
When a project targets multiple frameworks the
ResolvePackageDependenciesDesignTime MSBuild target has to be
called passing a TargetFramework property. When multiple target
frameworks are used by the project the MSBuild
ResolvePackageDependenciesDesignTime target is not available. So
the TargetFramework has to be specified so MSBuild treats the project
as targeting a single target framework. Note that this currently
only works if dotnet restore has been run on the project from the
command line. The NuGet package restore when multi-targeting is
currently not working. Also the project will show no files if it
targets multiple frameworks currently.