[Core] Handle msbuild items defined inside a target element
Fixes VSTS #528487 "The type or namespace name could not be found"
error thrown by Intellisense but compilation is ok.

If an MSBuild targets file defines an MSBuild Compile item inside a
Target element then this item would not be available to the type
system and would result in the text editor showing an error whilst
compiling the project would succeed.

The problem was that an ItemGroup inside a Target element was not
being evaluated.