[Core] Out of date project builder used on reload
Editing a .csproj file to change a reference and after the project
was reloaded the type system would still be using the old reference
information. The problem was that the type system service reloads
the project when the reloaded project is added to the solution.
This can happen before the old project is disposed, which shuts down
the project builder, resulting in the old out of date project builder
being used. To try to prevent this the project is unloaded from the
remote build engine before the SolutionItemAdded event is fired.
A new project builder should then be created when the type system
gets the reference information for the project.

Fixes VSTS #738993 - Type system not updated after saving edited
project file in editor