[Ide] Fix incorrect preprocessor symbols on switching framework
Switching to a different target framework in the text editor would
not update the preprocessor symbols available. The problem was that
the framework specific values were not being used when creating the
Roslyn project. Now a framework specific project configuration
is obtained from the project so the correct compiler options and
parse options are used by Roslyn.