[WIP][NuGet] Update multiple packages together in a batch
Updating two NuGet package references that have a strict dependency
on a single version of another NuGet package not explicitly added
to the project would fail in the Manage NuGet Packages dialog. This
was because the update would be done one package at a time. Doing
the update of the two NuGet packages together allows NuGet to
update both so they then use the new strict dependency.

Note that updating all packages in the project from the Solution
window does not have this problem since there the packages are
updated together in a batch.

Fixes VSTS #986960 - Solution nuget updater fails to update a set of
packages that all have a related dependency update