[Core] Support generating a binlog for any MSBuild target
Previously a binlog file would only be generated if a Build was run.
Now it is possible to have a binlog generated for any MSBuild target.
The filename for the bin log can be specified when running an MSBuild
target by setting the TargetEvaluationContext's BinLogFilePath.

Note that setting the BinLogFilePath if a build session is in progress
will not have any affect, the build session's binlog overrides this
setting.